SellerService

+++ src/main/java/com/crud/invoices/service/PaymentService.java
@@ -3,7 +3,7 @@ package com.crud.invoices.service;
 import com.crud.invoices.domain.Invoice;
 import com.crud.invoices.domain.InvoiceStatus;
 import com.crud.invoices.domain.Payment;
-import com.crud.invoices.exception.InvoiceNullPointerException;
+import com.crud.invoices.domain.PaymentStatus;
 import com.crud.invoices.respository.PaymentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -21,11 +21,33 @@ public class PaymentService {
     @Autowired
     InvoiceService invoiceService;
 
-    public Payment savePayment(Payment payment, final Long invoiceId) {
+    public Payment savePayment(Payment payment, final Long invoiceId) {///zmien
+        Optional<Invoice> invoiceInDatabase = invoiceService.getInvoice(invoiceId);
+        InvoiceStatus isSettled = invoiceInDatabase.get().getIsSettled();
+        //if (isSettled.equals(InvoiceStatus.FALSE)) {
+            //invoiceInDatabase.get().setIsSettled(InvoiceStatus.TRUE);
+        //}
+        BigDecimal currentPayment = payment.getPaid();
+        payment = new Payment(
+                null,
+                payment.getMethodOfPayment(),
+                currentPayment,
+                payment.getDateOfPayment(),
+                PaymentStatus.FALSE,
+                invoiceInDatabase.get()
+        );
+        this.addPaymentForExistingInvoice(invoiceInDatabase.get(), currentPayment);
+        paymentRepository.save(payment);
+        return payment;
+    }
+
+
+
+    public Payment settle(Payment payment, final Long invoiceId) {
         Optional<Invoice> invoiceInDatabase = invoiceService.getInvoice(invoiceId);
         InvoiceStatus isSettled = invoiceInDatabase.get().getIsSettled();
         if (isSettled.equals(InvoiceStatus.FALSE)) {
-            invoiceInDatabase.get().setIsSettled(InvoiceStatus.TRUE);
+        invoiceInDatabase.get().setIsSettled(InvoiceStatus.TRUE);
         }
         BigDecimal currentPayment = payment.getPaid();
         payment = new Payment(
@@ -33,16 +55,18 @@ public class PaymentService {
                 payment.getMethodOfPayment(),
                 currentPayment,
                 payment.getDateOfPayment(),
+                PaymentStatus.FALSE,
                 invoiceInDatabase.get()
         );
-        this.addPaymentForExistingSettledInvoice(invoiceInDatabase.get(),currentPayment);
-        return  paymentRepository.save(payment);
+        this.addPaymentForExistingInvoice(invoiceInDatabase.get(), currentPayment);
+        paymentRepository.save(payment);
+        return payment;
     }
 
 
 
     @Transactional
-    public void addPaymentForExistingSettledInvoice(Invoice invoice, BigDecimal currentPayment) {
+    public void addPaymentForExistingInvoice(Invoice invoice, BigDecimal currentPayment) {
         Optional<BigDecimal> sumOfPayments = Optional.of(
                 paymentRepository.findAllPaymentsByInvoiceId(invoice.getId())
                         .stream()
@@ -63,9 +87,6 @@ public class PaymentService {
         return paymentRepository.findAllPaymentsByInvoiceId(invoiceId);
     }
 
-    public Optional<Payment> findFirstPaymentByInvoiceId(final Long invoiceId) {
-        return Optional.ofNullable(paymentRepository.findFirstByInvoiceId(invoiceId).orElseThrow(InvoiceNullPointerException::new));
-    }
 
     public List<Payment> getAllPayments() {
         List<Payment> payments = paymentRepository.findAll();
@@ -73,6 +94,22 @@ public class PaymentService {
     }
 
     public void deletePayment(Long paymentId) {
-        this.paymentRepository.deleteById(paymentId);
+       Optional<Payment> paymentOptional = paymentRepository.findById(paymentId);
+      if(paymentOptional.isPresent()) {
+          Optional<Invoice> invoiceOptional = Optional.ofNullable(paymentOptional.get().getInvoice());
+          if(invoiceOptional.isPresent()) {
+             BigDecimal sumOfPayments =  invoiceOptional.get().getAmountPaid();
+              BigDecimal newSumOfPayments = sumOfPayments.subtract(paymentOptional.get().getPaid());
+              invoiceOptional.get().setAmountPaid(newSumOfPayments);
+              BigDecimal newleftToPay = invoiceOptional.get().getSumTotal().subtract(newSumOfPayments);
+              invoiceOptional.get().setLeftToPay(newleftToPay);
+              paymentRepository.save(paymentOptional.get());
+          }
+          this.paymentRepository.deleteById(paymentId);
+      }
     }
+
+
 }
+
+
